fix: Implementation of loading user into the session
