+++ src/context/UserContext.tsx
@@ -1,6 +1,7 @@
-import { ReactNode, useState } from "react";
+import { ReactNode, useEffect, useState } from "react";
 import { User } from "@/types";
 import { UserContext } from "./useUserContext";
+import { fetchUser } from "@/api/user";
 
 interface UserContextProviderProps {
   children: ReactNode;
@@ -8,6 +9,20 @@ interface UserContextProviderProps {
 
 function UserContextProvider({ children }: UserContextProviderProps) {
   const [user, setUser] = useState<User | null>(null);
+
+  useEffect(() => {
+    const userFromDb = fetchUser();
+    if (userFromDb) {
+      setUser({
+        _id: "123",
+        username: "HASOS",
+        fullname: "Itay Hasson",
+        email: "123@123.123",
+        homeCity: "Tel Aviv",
+        token: "123",
+      });
+    }
+  }, []);
 
   return (
     <UserContext.Provider value={{ user, setUser }}>
